perf: optimize codes
