+++ resources/assets/css/app.css
@@ -21,3 +21,4 @@
 @import "./components/_overlay";
 @import "./components/_countdown";
 @import "./components/_menu";
+@import "./components/_modal";
