+++ resources/assets/css/components/_modal.css
@@ -1,0 +1,192 @@
+/* ===== Order detail modal ===== */
+.order-modal {
+    position: fixed;
+    inset: 0;
+    z-index: 9999;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 1rem;
+}
+
+.order-modal[hidden] {
+    display: none;
+}
+
+.order-modal__backdrop {
+    position: absolute;
+    inset: 0;
+    background: rgba(15, 23, 42, 0.55);
+    backdrop-filter: blur(2px);
+}
+
+.order-modal__panel {
+    position: relative;
+    width: 100%;
+    max-width: 720px;
+    max-height: calc(100vh - 2rem);
+    overflow: hidden;
+    background: #fff;
+    border-radius: 1rem;
+    box-shadow: 0 24px 48px -12px rgba(0, 0, 0, 0.35);
+    animation: orderModalIn 180ms ease-out;
+}
+
+.order-modal__body {
+    max-height: calc(100vh - 2rem);
+    overflow-y: auto;
+}
+
+.order-modal__close {
+    position: absolute;
+    top: 0.875rem;
+    right: 0.875rem;
+    z-index: 10;
+    width: 34px;
+    height: 34px;
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    color: #fff;
+    font-size: 20px;
+    line-height: 1;
+    background: rgba(0, 0, 0, 0.18);
+    border: 1px solid rgba(255, 255, 255, 0.25);
+    border-radius: 9999px;
+    backdrop-filter: blur(8px);
+    cursor: pointer;
+    transition: background 150ms;
+}
+
+.order-modal__close:hover {
+    background: rgba(0, 0, 0, 0.35);
+}
+
+@keyframes orderModalIn {
+    from { opacity: 0; transform: translateY(8px) scale(0.98); }
+    to   { opacity: 1; transform: translateY(0) scale(1); }
+}
+
+/* ===== Order detail popup (SweetAlert2 override) ===== */
+.orders-detail-popup .swal2-title {
+    display: none !important;
+}
+
+.orders-detail-popup .swal2-html-container {
+    margin: 0 !important;
+    padding: 0 !important;
+    overflow: visible !important;
+    text-align: left !important;
+}
+
+.orders-detail-popup .swal2-close {
+    @apply z-50 text-white;
+    color: #fff !important;
+    top: 0.875rem !important;
+    right: 0.875rem !important;
+    background: rgba(0, 0, 0, 0.18) !important;
+    backdrop-filter: blur(8px);
+    border: 1px solid rgba(255, 255, 255, 0.25) !important;
+    border-radius: 9999px !important;
+    width: 34px !important;
+    height: 34px !important;
+    line-height: 1 !important;
+    font-size: 20px !important;
+    transition: background 150ms !important;
+}
+
+.orders-detail-popup .swal2-close:hover {
+    background: rgba(0, 0, 0, 0.35) !important;
+    color: #fff !important;
+    transform: none !important;
+}
+
+/* ===== Detail components ===== */
+.detail-modal {
+    @apply flex flex-col;
+}
+
+.detail-modal > *:not(.detail-hero) {
+    margin-left: 1.25rem;
+    margin-right: 1.25rem;
+}
+
+.detail-modal > *:last-child {
+    margin-bottom: 1.25rem;
+}
+
+.detail-hero {
+    @apply relative overflow-hidden px-5 pb-5 pt-6 text-white;
+    background: linear-gradient(135deg, #d84315 0%, #f57c00 60%, #ff7043 100%);
+}
+
+.detail-hero__bg::before,
+.detail-hero__bg::after {
+    content: '';
+    position: absolute;
+    border-radius: 9999px;
+    background: rgba(255, 255, 255, 0.06);
+    pointer-events: none;
+}
+
+.detail-hero__bg::before {
+    width: 220px;
+    height: 220px;
+    top: -110px;
+    right: -60px;
+}
+
+.detail-hero__bg::after {
+    width: 160px;
+    height: 160px;
+    bottom: -90px;
+    left: -40px;
+}
+
+.detail-chip {
+    @apply inline-flex items-center gap-1.5 rounded-full border px-2.5 py-1 text-[11px] font-bold transition;
+}
+
+.detail-chip--glass {
+    @apply border-white/30 bg-white/15 text-white backdrop-blur-md;
+}
+
+.detail-chip--status {
+    @apply border-transparent;
+}
+
+.detail-chip--primary {
+    @apply border-primary/30 bg-primary-softer text-primary hover:bg-primary-soft;
+}
+
+.detail-chip--success {
+    @apply border-emerald-200 bg-emerald-50 text-emerald-700 hover:bg-emerald-100;
+}
+
+.detail-chip--muted {
+    @apply border-mercury bg-skull text-onyx;
+}
+
+.detail-action {
+    @apply inline-flex flex-col items-center justify-center gap-1 rounded-xl border border-mercury bg-white px-3 py-3 text-[11px] font-semibold text-foreground transition hover:border-primary hover:bg-primary-softer hover:text-primary;
+}
+
+.detail-action > i {
+    @apply text-lg;
+}
+
+.detail-action--success {
+    @apply hover:border-emerald-400 hover:bg-emerald-50 hover:text-emerald-700;
+}
+
+.detail-card {
+    @apply rounded-2xl border border-mercury bg-white p-4;
+}
+
+.detail-card__title {
+    @apply mb-3 inline-flex items-center gap-1.5 text-[12px] font-semibold text-foreground;
+}
+
+.detail-card__title > i {
+    @apply text-base text-onyx;
+}
